# cloudify/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cloudify API
  description: |
    Cloudify is a self-hosted deployment platform for modern web applications.
    This API provides programmatic access to manage projects, deployments, domains,
    environment variables, serverless functions, and more.

    ## Authentication
    All API requests require authentication using a Bearer token.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Rate Limiting
    API requests are rate limited to 1000 requests per minute per API key.

    ## Errors
    The API uses standard HTTP response codes to indicate success or failure.
    - 2xx: Success
    - 4xx: Client error (bad request, unauthorized, not found)
    - 5xx: Server error
  version: 1.0.0
  contact:
    name: Cloudify Support
    email: support@cloudify.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://cloudify.tranthachnguyen.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /deployments/{deploymentId}:
    get:
      tags: [Deployments]
      summary: Get deployment details
      operationId: getDeployment
      parameters:
        - $ref: '#/components/parameters/DeploymentId'
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
...
  /ai/performance:
    post:
      tags: [AI]
      summary: Performance recommendations
      description: Get AI-powered performance recommendations for a project
      operationId: performanceRecommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId]
              properties:
                projectId:
                  type: string
      responses:
        '200':
          description: Performance recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendations:
                    type: array
                    items:
                      $ref: '#/components/schemas/PerformanceRecommendation'

  # ==================== Audit Logs ====================
  /audit-logs:
    get:
      tags: [Audit Logs]
      summary: List audit logs
      operationId: listAuditLogs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: type
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: string
        - name: teamId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /audit-logs/export:
    post:
      tags: [Audit Logs]
      summary: Export audit logs
      operationId: exportAuditLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                format:
                  type: string
                  enum: [json, csv]
                  default: json
                filters:
                  type: object
                  properties:
                    type:
                      type: string
                    action:
                      type: string
                    startDate:
                      type: string
                      format: date-time
                    endDate:
                      type: string
                      format: date-time
      responses:
        '200':
          description: Exported file
          content:
            application/json:
              schema:
                type: string
            text/csv:
              schema:
                type: string

  # ==================== Webhooks ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DeploymentId:
      name: deploymentId
      in: path
      required: true
      schema:
        type: string
      description: Deployment ID
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
...
    Deployment:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        status:
          type: string
          enum: [QUEUED, BUILDING, DEPLOYING, READY, ERROR, CANCELLED]
        branch:
          type: string
        commitSha:
          type: string
        commitMessage:
          type: string
        url:
          type: string
        isPreview:
          type: boolean
        duration:
          type: integer
          description: Build duration in milliseconds
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
    PerformanceRecommendation:
      type: object
      properties:
        category:
          type: string
          enum: [lcp, fid, cls, ttfb, bundle]
        severity:
          type: string
          enum: [low, medium, high]
        issue:
          type: string
        suggestion:
          type: string
        codeExample:
          type: string
        impact:
          type: string
    AuditLog:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        action:
          type: string
        description:
          type: string
        timestamp:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
        project:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
            slug:
              type: string
        team:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
            slug:
              type: string
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
security:
  - bearerAuth: []